+++ src/index.ts
@@ -3,11 +3,14 @@
 // OSU package entrypoint, for exporting all and utilities.
 
 // OSU name-dot-number and email utility exports
-export * from "@src/utils/validators";
+export * from "@src/validators/validators";
 
 // OSU fetch and data utility exports, with corresponding type exports.
-export * from "@src/utils/getStudentOrgs";
-export type { StudentOrg } from "@src/schemas/StudentOrg";
+export * from "@src/schemas/StudentOrg";
+export * from "@src/studentOrgs/getStudentOrgs";
 
-export * from "@src/utils/getUndergradMajors";
-export type { UndergradMajor } from "@src/schemas/UndergradMajor";
+export * from "@src/schemas/UndergradMajor";
+export * from "@src/undergradMajors/getUndergradMajors";
+
+export * from "@src/schemas/ResearchPosting";
+export * from "@src/researchPostings/getResearchPostingsAsync";