+++ src/undergradMajors/getUndergradMajors.test.ts
@@ -1,6 +1,6 @@
-// ./src/utils/getUndergradMajors.test.ts
+// ./src/undergradMajors/getUndergradMajors.test.ts
 //
-// Unittests for the getUndergradMajors utility module.
+// Unittests for the getUndergradMajors module.
 
 // Vitest essential imports.
 import { describe, expect, it } from "vitest";
@@ -11,8 +11,8 @@ import {
   UndergradMajorSchema,
 } from "@src/schemas/UndergradMajor";
 
-// Testing module imports.
-import { getUndergradMajors } from "@src/utils/getUndergradMajors";
+// Testing target import.
+import { getUndergradMajors } from "@src/undergradMajors/getUndergradMajors";
 
 // -----------------------------------------------------------------------------
 // Test suite for the getUndergradMajors utility function.